Update conditions for pirogue_admin.system_config changes.
It's not entirely clear at this point whether they belong to
package_config or system_config… They're useful in package_config but
system_config is really the one knowing about its own inner workings.

At this point we're missing the “this condition requires this and that
variable” information. While SYSTEM_OPERATING_MODE is a must, things
like SYSTEM_DNSMASQ might not exist. One way to implement that might be
to extend the decorator so that it can store both the function and a
list of required variables.

Additionally, iptables/nftables detection would make sense alongside
other network-related detection function in system_config.