Switch from a SYSTEM_DNSMASQ condition to an ENABLE_DHCP variable.
Rename and update the formatter to match the new type (since dhcp_range
became dhcp_range_line): we want to be able to comment it out entirely
when ENABLE_DHCP is false.

We don't ignore ENABLE_DHCP in AP mode, even if it's the mode where we
would expect DHCP to be needed: advanced users might have devices with
a manual IP configuration…

Set ENABLE_DHCP to true during autodetection as that ensures devices on
the isolated network have a chance of getting an IP address, and let
users switch it to false if they so wish.