Add a bool formatter.
At the moment we're only declaring ENABLE_DHCP in pirogue-networking's
admin section to reflect the fact we need it to come up with the value
with which @DHCP_RANGE_LINE@ must be replaced.

Even if we don't use @ENABLE_DHCP@ anywhere, the token management logic
requires strings to replace tokens with, so we can't stick to booleans.

Use str() as a formatter, entirely arbitrarily.