Add a safeguard around PUBLIC_EXTERNAL_NETWORK_ADDR.
Alternatively we could introduce a check_variables_consistency() method
that would be called before apply_configuration(). That would be a no-op
(for the time being) for PackageConfig objects, and only be implemented
by SystemConfig.

Alternatively, since we *absolutely* have to set the operating mode (to
something supported), it could be passed along as a parameter to the
get_needed_variables() methods; or at least to the SystemConfig one.