Bridge the gap between both CLI tools.
We could fold pirogue-admin-wireguard into pirogue-admin, but that'd
require some work, and proper integration is pending via a pirogue-admin
server exposing an API through gRPC, via a remote API, and via a local
pirogue-admin client.

Let's not spend more time on this right now, and re-use pirogue-admin
variables if any, parsing pirogue-admin's config.yaml file manually,
and keeping fallback variables if the file is missing or is missing
some variables.

This should make it possible for pirogue-base.postinst to leverage
pirogue-admin for any of the three supported modes, and for users to
manipulate peers using pirogue-admin-wireguard.