Clarify WireGuard is working with the public-facing external address.
We're not using any EXTERNAL_ADDRESS variable at this point, but one
could easily lose track of the fact the particular external address
we're interested in in the WireGuard case is the public one. Let's be
explicit.

This is mainly for clarity and consistency, no functional changes for
fresh deployments (the configuration file format gets a renamed key
though).