Use the comment to generate peer config/archive/QR code when needed.
Make it possible to specify a desired destination filename, but also
support getting a None parameter. In that case, build a filename
depending on the desired file extension and either the comment (if
non-empty) or the default format string, using the peer index.

While doing so, adjust the pirogue-admin-wireguard CLI tool to let
WgManager pick the filename.

Link: https://github.com/PiRogueToolSuite/pirogue-admin/issues/14