Add support for Wi-Fi interfaces in the NM case.
It seems NetworkManager doesn't want to let us set up just an IP address
as we would do in the ifupdown case, if the interface is configured with
mode=ap.

Go with the flow, and include as many settings as needed to make the
connection usable from a test phone.

Link: https://github.com/PiRogueToolSuite/pirogue-admin/issues/8