Improve interface detection.
Do that by making sure each /sys/class/net entry features a `type` (see
Documentation/ABI/testing/sysfs-class-net). This fixes autodetection
when the `bonding` module is loaded, as it maintains a regular file
called `bonding_masters` that would otherwise be considered a valid
candidate as an isolated interface. With thanks to the Qurium team for
the report!

Link: https://github.com/PiRogueToolSuite/pirogue-admin/issues/25